requirements/local_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=local_requirements.txt base.in local.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
click==7.0                # via pip-tools
colorama==0.4.1           # via pytest
coverage==4.5.3           # via pytest-cov
django-debug-toolbar==1.9
django-widget-tweaks==1.4.5
django==1.10.5
factory-boy==2.11.0
faker==1.0.7              # via factory-boy
importlib-metadata==0.18  # via pluggy, pytest
markdown==3.1.1
more-itertools==7.1.0     # via pytest
packaging==19.0           # via pytest
pip-tools==3.8.0
pluggy==0.12.0            # via pytest
psycopg2==2.8.3
py==1.8.0                 # via pytest
pyparsing==2.4.0          # via packaging
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==5.0.0
python-dateutil==2.8.0    # via faker
six==1.12.0               # via faker, packaging, pip-tools, python-dateutil
sqlparse==0.3.0           # via django-debug-toolbar
text-unidecode==1.2       # via faker
wcwidth==0.1.7            # via pytest
zipp==0.5.1               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via markdown
